update brightness in response to scroll wheel
Rather than specifically connecting to the 'drag-end' events, we update
the brightness whenever the slider value is changed. We already have
debouncing to avoid changing the brightness too often, so calling
"_broadcastBrightness" frequently is harmless.

This change has the addition positive effect that the brightness is
changed while the user is dragging if they hover on a particular value
for long enough. This is convenient for precise brightness selection.